Move DummyProtocolsHandler out of the module
